requirements.txt
# Core data science and analysis
pandas>=1.5.0
numpy>=1.20.0

# Technical analysis
TA-Lib>=0.4.21

# Market data APIs
alpaca-trade-api>=3.0.0

# Environment and configuration
python-dotenv>=0.19.0

# Web requests
requests>=2.28.0

# Timezone handling
pytz>=2022.1

# Additional utility packages
pathlib2>=2.3.7  # For older Python compatibility if needed